requirements.txt
praw
discord.py
tinydb
python-dotenv